fix model

+++ include/model.hpp
@@ -17,7 +17,7 @@ public:
 		assert(0 < R && R <= 250);
 		int buf_size = (R*R*R+7)/8;
 		buf = new char[buf_size];
-		if (fread(buf, 1, buf_size, fp) != buf_size) {
+		if ((int)fread(buf, 1, buf_size, fp) != buf_size) {
 			throw "Model fread failure";
 		}
 		fclose(fp);
